fix: app_state.g

+++ core/app_state.h
@@ -41,11 +41,11 @@ struct ErrorState {
 struct PacNVMem
 {
     uint8_t *nr_fixnv1_mem = nullptr;
-    size_t nr_fixnv1_mem_size = 0;
+    uint64_t nr_fixnv1_mem_size = 0;
     uint8_t *l_fixnv1_mem = nullptr;
-    size_t l_fixnv1_mem_size = 0;
+    uint64_t l_fixnv1_mem_size = 0;
     uint8_t *downloadnv_mem = nullptr;
-    size_t downloadnv_mem_size = 0;
+    uint64_t downloadnv_mem_size = 0;
 };
 
 // 集中管理应用运行时状态，替代分散的 extern 全局变量
